requirements.txt
nicegui
psycopg[binary]
python-dotenv
reportlab
requests
fastapi
uvicorn
openpyxl
